chore: Add Middleware in store redux for DeleteUsers
This commit adds the middleware and Toaster component from the 'sonner' library to the App component in App.tsx. It also updates the dependencies in package.json and package-lock.json to include the 'sonner' package with version 1.5.0. This allows for displaying rich-colored toasts in the application. Additionally, the commit updates the types in types.d.ts to use FormEvent instead of ChangeEvent for EventsChange and EventsSubmit. Lastly, a new file useUserActions.ts is added to handle user actions in the application.